Added db context design factory
