+++ Backend/CarRentalApp/CarRentalApp.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31919.166
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CarRentalApp", "CarRentalApp\CarRentalApp.csproj", "{1886D836-30ED-4896-B26D-4C1769FE8BCD}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CarRentalWeb", "CarRentalWeb\CarRentalWeb.csproj", "{1886D836-30ED-4896-B26D-4C1769FE8BCD}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CarRentalDal", "CarRentalDal\CarRentalDal.csproj", "{F57B80FE-120B-493B-8BE0-1A2D8D416D6B}"
 EndProject
