+++ Backend/CarRentalApp/CarRentalBll/CarRentalBll.csproj
@@ -7,11 +7,8 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\CarRentalDal\CarRentalDal.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <PackageReference Include="Mapster" Version="7.3.0" />
+        <ProjectReference Include="..\CarRentalDal\CarRentalDal.csproj"/>
+        <PackageReference Include="Mapster" Version="7.3.0"/>
     </ItemGroup>
 
 </Project>
