+++ Backend/CarRentalApp/CarRentalWeb/Controllers/UsersController.cs
@@ -1,11 +1,11 @@
 ﻿using System.Collections.Immutable;
-using CarRentalApp.Models.Web.Requests;
 using CarRentalBll.Models;
 using CarRentalBll.Services;
+using CarRentalWeb.Models.Requests;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
-namespace CarRentalApp.Controllers
+namespace CarRentalWeb.Controllers
 {
     [ApiController]
     [Route("api/[controller]")]
