+++ Backend/CarRentalApp/CarRentalWeb/appsettings.json
@@ -42,7 +42,6 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "CarRentalDB1": "Server=WSB-084-74\\MSSQSERVER1;Database=CarRentalDB;Trusted_Connection=True;",
-    "CarRentalDB": "Server=DESKTOP-I1L3GQA\\MSSQLSERVER01;Database=CarRentalDB;Trusted_Connection=True;"
+    "CarRentalDB": "Server=WSB-084-74\\MSSQSERVER1;Database=CarRentalDB;Trusted_Connection=True;"
   }
 }
